# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Light For Change - Digital Candle Map API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /candles:
    get:
      operationId: getCandles
      tags: [candles]
      summary: List candles with optional viewport bounds
      parameters:
        - name: minLat
          in: query
          schema: { type: number }
        - name: maxLat
          in: query
          schema: { type: number }
        - name: minLng
          in: query
          schema: { type: number }
        - name: maxLng
          in: query
          schema: { type: number }
        - name: country
          in: query
          schema: { type: string }
        - name: state
          in: query
          schema: { type: string }
        - name: city
          in: query
          schema: { type: string }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 500 }
        - name: fromDate
          in: query
          schema: { type: string, format: date-time }
        - name: toDate
          in: query
          schema: { type: string, format: date-time }
      responses:
        "200":
          description: List of candles
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CandleList"

    post:
      operationId: createCandle
      tags: [candles]
      summary: Light a candle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CandleInput"
      responses:
        "201":
          description: Candle created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Candle"
        "429":
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /candles/recent:
    get:
      operationId: getRecentCandles
      tags: [candles]
      summary: Most recent candles with messages
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
      responses:
        "200":
          description: Recent candles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Candle"
  /candles/clusters:
    get:
      operationId: getCandleClusters
      tags: [candles]
      summary: Clustered candle data for map rendering
      parameters:
        - name: zoom
          in: query
          required: true
          schema: { type: number }
        - name: minLat
          in: query
          schema: { type: number }
        - name: maxLat
          in: query
          schema: { type: number }
        - name: minLng
          in: query
          schema: { type: number }
        - name: maxLng
          in: query
          schema: { type: number }
      responses:
        "200":
          description: Cluster data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CandleCluster"
  /candles/timeline:
    get:
      operationId: getCandleTimeline
      tags: [candles]
      summary: Daily candle counts for timeline slider
      parameters:
        - name: days
          in: query
          schema: { type: integer, default: 30 }
      responses:
        "200":
          description: Timeline data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TimelineEntry"
...
  /locations/states:
    get:
      operationId: getStates
      tags: [locations]
      summary: List of states for a country
      parameters:
        - name: country
          in: query
          required: true
          schema: { type: string }
      responses:
        "200":
          description: States
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
  /locations/cities:
    get:
      operationId: getCities
      tags: [locations]
      summary: List of cities for a state
      parameters:
        - name: state
          in: query
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Cities
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    Candle:
      type: object
      properties:
        id:
          type: integer
        displayName:
          type: ["string", "null"]
        message:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        state:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        latitudeRounded:
          type: ["number", "null"]
        longitudeRounded:
          type: ["number", "null"]
        createdAt:
          type: string
          format: date-time
        isHidden:
          type: boolean
      required: [id, createdAt, isHidden]
    CandleInput:
      type: object
      properties:
        displayName:
          type: string
          maxLength: 50
        message:
          type: string
          maxLength: 150
        country:
          type: string
        state:
          type: string
        city:
          type: string
        latitude:
          type: number
          minimum: -90
          maximum: 90
        longitude:
          type: number
          minimum: -180
          maximum: 180
        captchaToken:
          type: string
        acceptedTerms:
          type: boolean
          description: Must be true to confirm acceptance of the published terms and community rules.
      required: [acceptedTerms]
    CandleList:
      type: object
      properties:
        candles:
          type: array
          items:
            $ref: "#/components/schemas/Candle"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [candles, total, page, limit]
    CandleCluster:
      type: object
      properties:
        lat:
          type: number
        lng:
          type: number
        count:
          type: integer
        isCluster:
          type: boolean
        candleId:
          type: ["integer", "null"]
        displayName:
          type: ["string", "null"]
        message:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        state:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        createdAt:
          type: ["string", "null"]
      required: [lat, lng, count, isCluster]
    TimelineEntry:
      type: object
      properties:
        date:
          type: string
          format: date
        count:
          type: integer
        cumulative:
          type: integer
      required: [date, count, cumulative]
...