# lib/api-zod/src/generated/types/candleCluster.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Light For Change - Digital Candle Map API
 * OpenAPI spec version: 0.1.0
 */

export interface CandleCluster {
  lat: number;
  lng: number;
  count: number;
  isCluster: boolean;
  /** @nullable */
  candleId?: number | null;
  /** @nullable */
  displayName?: string | null;
  /** @nullable */
  message?: string | null;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  state?: string | null;
  /** @nullable */
  country?: string | null;
  /** @nullable */
  createdAt?: string | null;
}